requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
itsdangerous==2.1.2


scikit-build 
setuptools
sse-starlette
pydantic-settings
starlette-context

haystack-ai


tqdm
PyMuPDF==1.26.1


# websockets==12.0

pdfminer.six

pymilvus 
milvus-haystack
fastembed-haystack

sentence-transformers==4.1.0

transformers[torch,sentencepiece]


